requirements.txt
openai
ollama
kokoro
requests
jsonschema
pyaudio
gtts
playsound
python-dotenv
soundfile
sounddevice
SpeechRecognition
mcp